refactor - Dev: normalize project
(bonus)
