+++ library/backlight/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/backlight/Greeting.kt
@@ -1,0 +1,6 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.backlight
+
+expect class Greeting() {
+
+    fun hello(): String
+}
