--- library/backlight/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/command/backlight/Greeting.kt
@@ -1,6 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.backlight
-
-expect class Greeting() {
-
-    fun hello(): String
-}
