+++ library/backlight/src/nativeMain/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/backlight/Greeting.kt
@@ -1,0 +1,7 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.backlight
+
+actual class Greeting {
+    private val helloWorld: String = "Hello Native!"
+
+    actual fun hello() = helloWorld
+}
