--- library/backlight/src/nativeMain/kotlin/io/github/gouvinb/wwmaccompanist/command/backlight/Greeting.kt
@@ -1,7 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.backlight
-
-actual class Greeting {
-    private val helloWorld: String = "Hello Native!"
-
-    actual fun hello() = helloWorld
-}
