--- library/backlight/src/nativeTest/kotlin/io/github/gouvinb/wwmaccompanist/command/backlight/GreetingTest.kt
@@ -1,13 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.backlight
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class GreetingTest : GreetingTestCommon() {
-
-    @Test
-    override fun hello() {
-        super.hello()
-        assertTrue("Native" in greeting.hello())
-    }
-}
