+++ library/bar/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/bar/Greeting.kt
@@ -1,0 +1,6 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.bar
+
+expect class Greeting() {
+
+    fun hello(): String
+}
