--- library/bar/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/command/bar/Greeting.kt
@@ -1,6 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.bar
-
-expect class Greeting() {
-
-    fun hello(): String
-}
