+++ library/bar/src/commonTest/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/bar/GreetingTestCommon.kt
@@ -1,0 +1,14 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.bar
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+open class GreetingTestCommon {
+
+    internal val greeting = Greeting()
+
+    @Test
+    open fun hello() {
+        assertTrue(greeting.hello() matches """Hello [A-Za-z]+!""".toRegex())
+    }
+}
