+++ library/bar/src/nativeMain/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/bar/Greeting.kt
@@ -1,0 +1,7 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.bar
+
+actual class Greeting {
+    private val helloWorld: String = "Hello Native!"
+
+    actual fun hello() = helloWorld
+}
