--- library/bar/src/nativeMain/kotlin/io/github/gouvinb/wwmaccompanist/command/bar/Greeting.kt
@@ -1,7 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.bar
-
-actual class Greeting {
-    private val helloWorld: String = "Hello Native!"
-
-    actual fun hello() = helloWorld
-}
