+++ library/launcher/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/launcher/Greeting.kt
@@ -1,0 +1,6 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.launcher
+
+expect class Greeting() {
+
+    fun hello(): String
+}
