--- library/launcher/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/command/launcher/Greeting.kt
@@ -1,6 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.launcher
-
-expect class Greeting() {
-
-    fun hello(): String
-}
