+++ library/launcher/src/nativeMain/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/launcher/Greeting.kt
@@ -1,0 +1,7 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.launcher
+
+actual class Greeting {
+    private val helloWorld: String = "Hello Native!"
+
+    actual fun hello() = helloWorld
+}
