--- library/launcher/src/nativeMain/kotlin/io/github/gouvinb/wwmaccompanist/command/launcher/Greeting.kt
@@ -1,7 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.launcher
-
-actual class Greeting {
-    private val helloWorld: String = "Hello Native!"
-
-    actual fun hello() = helloWorld
-}
