+++ library/screenshot/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/screenshot/Greeting.kt
@@ -1,0 +1,6 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.screenshot
+
+expect class Greeting() {
+
+    fun hello(): String
+}
