--- library/screenshot/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/command/screenshot/Greeting.kt
@@ -1,6 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.screenshot
-
-expect class Greeting() {
-
-    fun hello(): String
-}
