+++ library/screenshot/src/nativeMain/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/screenshot/Greeting.kt
@@ -1,0 +1,7 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.screenshot
+
+actual class Greeting {
+    private val helloWorld: String = "Hello Native!"
+
+    actual fun hello() = helloWorld
+}
