--- library/screenshot/src/nativeMain/kotlin/io/github/gouvinb/wwmaccompanist/command/screenshot/Greeting.kt
@@ -1,7 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.screenshot
-
-actual class Greeting {
-    private val helloWorld: String = "Hello Native!"
-
-    actual fun hello() = helloWorld
-}
