+++ library/theme/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/theme/Greeting.kt
@@ -1,0 +1,6 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.theme
+
+expect class Greeting() {
+
+    fun hello(): String
+}
