--- library/theme/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/command/theme/Greeting.kt
@@ -1,6 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.theme
-
-expect class Greeting() {
-
-    fun hello(): String
-}
