--- library/theme/src/commonTest/kotlin/io/github/gouvinb/wwmaccompanist/command/theme/GreetingTestCommon.kt
@@ -1,14 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.theme
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-open class GreetingTestCommon {
-
-    internal val greeting = Greeting()
-
-    @Test
-    open fun hello() {
-        assertTrue(greeting.hello() matches """Hello [A-Za-z]+!""".toRegex())
-    }
-}
