+++ library/wallpaper/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/wallpaper/Greeting.kt
@@ -1,0 +1,6 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.wallpaper
+
+expect class Greeting() {
+
+    fun hello(): String
+}
