--- library/wallpaper/src/commonMain/kotlin/io/github/gouvinb/wwmaccompanist/command/wallpaper/Greeting.kt
@@ -1,6 +1,0 @@
-package io.github.gouvinb.wwmaccompanist.command.wallpaper
-
-expect class Greeting() {
-
-    fun hello(): String
-}
