+++ library/wallpaper/src/nativeMain/kotlin/io/github/gouvinb/wwmaccompanist/audio/command/wallpaper/Greeting.kt
@@ -1,0 +1,7 @@
+package io.github.gouvinb.wwmaccompanist.audio.command.wallpaper
+
+actual class Greeting {
+    private val helloWorld: String = "Hello Native!"
+
+    actual fun hello() = helloWorld
+}
